Rewrite interface for query service
